Move hooks to common if referenced from multiple places.
