+++ src/pages/workspaces/workspace/Dashboard/AzureStorageDetails.ts
@@ -6,7 +6,7 @@ import { ReactComponent as AzureLogo } from 'src/images/azure.svg';
 import { getRegionFlag, getRegionLabel } from 'src/libs/azure-utils';
 import { AzureContext } from 'src/libs/workspace-utils';
 import { InfoRow } from 'src/pages/workspaces/workspace/Dashboard/InfoRow';
-import { StorageDetails } from 'src/workspaces/container/state/useWorkspace';
+import { StorageDetails } from 'src/workspaces/common/state/useWorkspace';
 
 interface AzureStorageDetailsProps {
   azureContext: AzureContext;
